threshold_detection/requirements.txt
numpy==1.19.5
Pillow==9.0.1
matplotlib==3.5.1
scikit-image==0.19.2